Create Control.v
First draft of the control. In this design the human ultimately acts as the program counter and the memory. Currently has problems reading from and writing to the same register. (i.e read A and B and write A+B to C is fine, but read A and C and write A-C to C does not work correctly.)